more refactoring to make each step modular
