+++ topik/fileio/__init__.py
@@ -1,0 +1,8 @@
+# these imports register the functions with the
+# registered_tokenizers function registry.
+from .elasticsearch import elastic, ElasticsearchOutput
+from .solr import import_solr
+from .dictionary import DictionaryOutput
+from .json import json_stream, large_json
+
+from ._registry import registered_inputs, registered_outputs, read_input
