--- topik/fileio/vectorized_corpus.py
@@ -1,1 +1,0 @@
-# TODO: implement
