+++ topik/models/__init__.py
@@ -1,7 +1,6 @@
+# these imports register the functions with the
+# registered_models function registry.
 from .lda import LDA
 from .plsa import PLSA
 
-from .model_base import load_model
-from ._registered_models import ModelRegistry
-
-registered_models = ModelRegistry()
+from ._registry import registered_models, run_model
