+++ topik/models/lda.py
@@ -3,18 +3,17 @@ from __future__ import absolute_import
 import pandas as pd
 
 import gensim
-from topik.intermediaries.digested_document_collection import DigestedDocumentCollection
 from topik.fileio.tokenized_corpus import TokenizedCorpus
-from topik.fileio.raw_corpus import load_persisted_corpus
-from .model_base import TopicModelResultBase
-from ._registered_models import register_model
+from topik.fileio.base_output import load_persisted_corpus
+from .base_model_output import TopicModelResultBase
+from ._registry import register
 
 # Doctest imports
 from topik.fileio.readers import read_input
 from topik.tests import test_data_path
 
 
-@register_model
+@register
 class LDA(TopicModelResultBase):
     """A high-level interface for an LDA (Latent Dirichlet Allocation) model.
