+++ topik/tokenizers/__init__.py
@@ -1,1 +1,7 @@
-from tokenizers=_registry import tokenize
+# these imports register the functions with the
+# registered_tokenizers function registry.
+from .entities import entities
+from .simple import simple
+from .ngrams import ngrams
+
+from ._registry import registered_tokenizers, tokenize
