--- topik/tokenizers/tokenizers.py
@@ -1,12 +1,0 @@
-from __future__ import absolute_import, print_function
-
-import itertools
-import logging
-import re
-
-
-
-logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s',
-                    level=logging.INFO)
-
-
