+++ topik/visualizers/__init__.py
@@ -1,1 +1,6 @@
-__author__ = 'msarahan'
+# these imports register the functions with the
+# registered_tokenizers function registry.
+from .termite import termite
+from .pyldavis import lda_vis
+
+from ._registry import registered_visualizers, visualize
