+++ topik/visualizers/termite.py
@@ -4,9 +4,10 @@ import pandas as pd
 import bokeh.plotting as plt
 from bokeh.models.sources import ColumnDataSource
 
+from ._registry import register
 
-@register_visualizer
-def plot(input_data, plot_title="Termite plot", output_file="termite.html"):
+@register
+def termite(input_data, plot_title="Termite plot", output_file="termite.html"):
     """A Bokeh Termite Visualization for LDA results analysis.
 
     Parameters
@@ -25,7 +26,7 @@ def plot(input_data, plot_title="Termite plot", output_file="termite.html"):
 
     """
 
-    t = blz.Data(self.input_file)
+    t = blz.Data(input_data)
 
     MAX = blz.compute(t.weight.max())
     MIN = blz.compute(t.weight.min())
@@ -48,7 +49,7 @@ def plot(input_data, plot_title="Termite plot", output_file="termite.html"):
 
     p = plt.figure(x_range=TOPICS, y_range=WORDS,
                    plot_width=1000, plot_height=1700,
-                   title=self.title)
+                   title=plot_title)
 
     p.circle(x="topic", y="word", size="size", fill_alpha=0.6, source=data_source)
     plt.show(p)
